Лр 5
лррр 5

+++ delete_key.php
@@ -1,9 +1,13 @@
 <?php
+include("checks.php");
 require_once 'connect1.php';
 $mysqli = new mysqli($host, $user, $password, $database) or die ("Невозможно
 подключиться к серверу");
 $id = $_GET['id'];
-$result = $mysqli->query("DELETE FROM dkey WHERE id='$id'");
-header("Location: key.php");
+if ($_SESSION['type'] == 2)
+    $result = $mysqli->query("DELETE FROM dkey WHERE id='$id'");
+else
+    echo "Недостаточно прав";
+header("Location: keyAdm.php");
 exit;
 ?>
